+++ tests/test_managing_actor.py
@@ -177,9 +177,7 @@ async def test_simple_start_stop(
 
     event = test_env.actor(1).initial_dispatch
     assert event.options == {"test": True}
-    assert event.components == TargetIds(
-        ComponentId(1), ComponentId(10), ComponentId(15)
-    )
+    assert event.target == TargetIds(ComponentId(1), ComponentId(10), ComponentId(15))
     assert event.dry_run is False
 
     assert test_env.actor(1).is_running is True
@@ -309,7 +307,7 @@ async def test_dry_run(test_env: _TestEnv, fake_time: time_machine.Coordinates) 
     event = test_env.actor(1).initial_dispatch
 
     assert event.dry_run is dispatch.dry_run
-    assert event.components == dispatch.target
+    assert event.target == dispatch.target
     assert event.options == dispatch.payload
     assert test_env.actor(1).is_running is True